refactor(consts): update index to answer-status to better reading od file
