+++ packages/challenges/src/infra/Messaging/kafka-messaging.producer.ts
@@ -4,7 +4,7 @@ import { Inject, Injectable, OnModuleInit } from '@nestjs/common';
 import { ClientKafka } from '@nestjs/microservices';
 import { lastValueFrom } from 'rxjs';
 import { PrismaService } from '../database/prisma/prisma.service';
-import { ANSWER_STATUS } from '@/core/consts';
+import { ANSWER_STATUS } from '@/core/consts/answer-status';
 
 interface CorrectLessonResponse {
   grade: number;
