+++ packages/challenges/src/infra/http/gql/modules/Answer/resolvers/register-answer-enum.ts
@@ -1,4 +1,4 @@
-import { ANSWER_STATUS } from '@/core/consts';
+import { ANSWER_STATUS } from '@/core/consts/answer-status';
 import { registerEnumType } from '@nestjs/graphql';
 
 export enum GQL_ANSWER_STATUS_ENUM {
